+++ src/util/ccn-lite-pktdump.c
@@ -502,7 +502,7 @@ ccntlv_parse_sequence(int lev, unsigned char ctx, unsigned char *base,
             return -1;
         }
 
-        if (vallen > *len) {
+        if ((int) vallen > *len) {
           fprintf(stderr, "\n%04zx ** CCNTLV length problem:\n"
               "  type=0x%04hx, len=0x%04hx larger than %d available bytes\n",
               *buf - base, (unsigned short)typ, (unsigned short)vallen, *len);
@@ -851,7 +851,7 @@ cistlv_parse_sequence(int lev, unsigned char ctx, unsigned char *base,
             return -1;
         }
 
-        if (vallen > *len) {
+        if ((int) vallen > *len) {
           fprintf(stderr, "\n%04zx ** CISTLV length problem:\n"
               "  type=0x%04hx, len=0x%04hx larger than %d available bytes\n",
               *buf - base, (unsigned short)typ, (unsigned short)vallen, *len);
@@ -1113,7 +1113,7 @@ iottlv_parse_sequence(int lev, unsigned char ctx, unsigned char *base,
             return -1;
         }
 
-        if (vallen > *len) {
+        if ((int) vallen > *len) {
             fprintf(stderr, "\n%04zx ** IOTTLV length problem:\n"
               "  type=0x%04hx, len=0x%04hx larger than %d available bytes\n",
               *buf - base, (unsigned short)typ, (unsigned short)vallen, *len);
@@ -1442,7 +1442,7 @@ localrpc_parse(int lev, unsigned char *base, unsigned char **buf, int *len,
                 printf("  ");
 
             printf("\"%.*s\"", maxlen, *buf);
-            if (vallen > maxlen)
+            if (vallen > (int) maxlen)
                 printf("...");
             printf("\n");
         } else if (vallen > 0)
